Another intermediary commit, a few tests to fix still.
