+++ ocean_lib/ocean/ocean.py
@@ -28,7 +28,6 @@ from ocean_lib.ocean.util import (
     get_web3_connection_provider,
     to_base_18,
 )
-from ocean_lib.web3_internal.contract_handler import ContractHandler
 from ocean_lib.web3_internal.utils import get_network_name
 from ocean_lib.web3_internal.wallet import Wallet
 from ocean_lib.web3_internal.web3_provider import Web3Provider
